Add code-extract.
This commit adds the first iteration of code-extract for pulling
functions and their contexts into standalone files. At this point, we
are able to build dummy calls to the functions being "pulled" such that
the standalone file is compiled. A few known issues are captured below:

- We cannot pull base classes unless referenced in dependent code
directly.
- We do not support pulling of external functions (i.e., functions not
defined in the current project).
- We do not support pulling function template specializations or class
template specializations.